+++ index.html
@@ -41,6 +41,6 @@
         </ul>
       </div>
     </main>
-    <script src="assets/js/script.js"></script>
+    <script type="module" src="assets/js/script.js"></script>
   </body>
 </html>
